requirements/base.txt
aiohttp~=3.7
aiocontextvars~=0.2
aiotask-context~=0.6
async-timeout~=3.0
asyncpg~=0.21
elastic-apm~=6.7
multidict~=5.1
psutil~=5.8
python-json-logger~=2.0
setuptools~=65.3
ujson~=4.1
sanic~=21.12
tortoise-orm~=0.18
pydantic~=1.9
sentry-sdk~=1.7
GitPython~=3.1